+++ R/read_licor.R
@@ -16,10 +16,10 @@
 #' @export
 read_licor <- function(filepath, model) {
   if (model == "7000"){
-    df <- read_li7000(filepath)
+    df <- read_licor_li7000(filepath)
     ensure_column_types(df)
   } else if (model == "850"){
-    df <- read_li850(filepath)
+    df <- read_licor_li850(filepath)
     ensure_column_types(df)
   }
 
@@ -34,34 +34,4 @@ read_licor <- function(filepath, model) {
     dplyr::relocate("time", .before = 1)
 
   return(df)
-}
-
-#' Read from LI-7000
-#' @description
-#' `r lifecycle::badge('experimental')`
-#' @inheritParams read_licor
-#' @keywords internal
-read_li7000 <- function(filepath){
-  filepath |>
-    readr::read_tsv(skip = 2,
-             name_repair = janitor::make_clean_names,
-             show_col_types = FALSE) |>
-    dplyr::rename(time = "time_s")
-}
-
-#' Read from LI-850
-#' @description
-#' `r lifecycle::badge('experimental')`
-#' @inheritParams read_licor
-#' @importFrom hms as_hms
-#' @keywords internal
-read_li850 <- function(filepath){
-  filepath |>
-    readr::read_tsv(skip = 1,
-             name_repair = janitor::make_clean_names,
-             show_col_types = FALSE) |>
-    suppressWarnings() |>
-    dplyr::select(where(.not_any_na)) |>
-    dplyr::mutate(system_time_h_m_s = hms::as_hms(.data$system_time_h_m_s)) |>
-    dplyr::mutate(time = .data$system_time_h_m_s - min(.data$system_time_h_m_s))
 }